requirements.txt
pytorch<2.6
torchvision
pandas
matplotlib
scikit-learn
timm
h5py